Add files via upload
Guessing Game:
I developed a program that generates a random number and challenges the user to guess it. The user inputs their guess, and the program provides feedback if the guess is too high or too low, continuing until the user guesses correctly and displaying the number of attempts it took to win the game.